refactor: update file srtucture
