+++ src/api/index.ts
@@ -1,0 +1,4 @@
+import API from "./endpoints";
+export type { IPhoto } from "./types";
+export { getAuthCodeUrl, fetchToken } from "./auth";
+export { API };
